block read and write support added back in.
Right now we differentiate between absolute and relative addresses. This becomes an issue when dealing with a memory space container because we transform the address as part of the read/write process from absolute to relative. To solve this, I believe we must make the children of the memory space containers relative to the base address of the container itself. So the memory spaces inside the container are never aware of what the real parent address is.

This also has the positive side effect of making it easier to maintain a tree of memory mapped components